requirements.txt
aiohttp==3.9.1
gitpython==3.1.40
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
structlog==23.2.0
python-dateutil==2.8.2